Chores: the first significant commit.
The window now shows a view of a larger cairo image that is fixed in the background
Presently the background cairo surface is a grid of lines instead of the map downloaded from OpenStreetMaps
The main drawing area responds to button presses, button releases, mouse moves. With these three events
a mouse drag is simulated and the offset for viewing the big map area is moved.